+++ pystiche_papers/sanakoyeu_et_al_2018/__init__.py
@@ -3,4 +3,5 @@ from ._data import *
 from ._loss import *
 from ._modules import *
 from ._nst import *
+from ._transformer import *
 from ._utils import *
